修改考勤模块和student的vo，增加js排序，修改样式表。header修改密码可用。

+++ trunk/ClassMS/src/cn/gdpu/action/StudentAction.java
@@ -115,6 +115,9 @@ public class StudentAction extends ActionSupport implements RequestAware {
 	}
 	
 	public String deleteMany() {
+		if(stuIds == null || stuIds.length < 1) {
+			return "index";
+		}
 		for(int i=0;i<stuIds.length;i++) {
 			studentService.deleteById(stuIds[i]);
 		}
